+++ CHANGELOG.md
@@ -17,6 +17,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Throw an error for OPTIONS on nonexistent tables - @calebmer
 - Remove deadlock on simultaneous contentious updates - @ruslantalpa, @begriffs
 
+### Added
+- Applies range headers to RPC calls - @diogob
+
 ## [0.3.0.3] - 2016-01-08
 
 ### Fixed